Linux fixes

+++ src/common/Coding/GScriptVM.cpp
@@ -1972,7 +1972,7 @@ public:
 										pos = strstr(sStart, sub);
 
 									if (pos)
-										*Dst = pos - s;
+										*Dst = (int64) (pos - s);
 									else
 										*Dst = -1;
 									break;
